% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/histml.R
\name{histml}
\alias{histml}
\title{Detailed Histogram}
\usage{
histml(x, y = "Sturges")
}
\arguments{
\item{x}{A vector of values for which a histogram is desired}

\item{y}{}
}
\description{
Detailed histogram with mean, median and normal curve.
}
\examples{
histml(myData, "BFI")
}
\keyword{histogram}

