% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/revScore.R
\name{revScore}
\alias{revScore}
\title{Interactive Reverse Scoring Function}
\usage{
revScore()
}
\description{
An interactive function for quickly performing statistical analyses on a variety of subscales.  It is designed for researchers who need to document their analysis in a script file that can be easily shared.
}
\examples{
revScore()
}
\keyword{score}
\keyword{subscales}

